Let each type handle gathering its own props
